requirements.txt
line-bot-sdk==1.15.0
firebase-admin==3.2.0